moved joi separetly
