+++ schemas/index.js
@@ -1,0 +1,4 @@
+const contactAddSchema = require("./contactAddSchema");
+const contactPutSchema = require("./contactPutSchema");
+
+module.exports = { contactAddSchema, contactPutSchema };
